// packages/macro-ai-api-client/src/client/client/types.gen.ts
export interface Config<T extends ClientOptions = ClientOptions>
	extends Omit<CreateAxiosDefaults, 'auth' | 'baseURL' | 'headers' | 'method'>,
		CoreConfig {
	/**
	 * Axios implementation. You can use this option to provide either an
	 * `AxiosStatic` or an `AxiosInstance`.
	 *
	 * @default axios
	 */
	axios?: AxiosStatic | AxiosInstance
	/**
	 * Base URL for all requests made by this client.
	 */
	baseURL?: T['baseURL']
	/**
	 * An object containing any HTTP headers that you want to pre-populate your
	 * `Headers` object with.
	 *
	 * {@link https://developer.mozilla.org/docs/Web/API/Headers/Headers#init See more}
	 */
	headers?:
		| AxiosRequestHeaders
		| Record<
				string,
				| string
				| number
				| boolean
				| (string | number | boolean)[]
				| null
				| undefined
				| unknown
		  >
	/**
	 * Throw an error instead of returning it in the response?
	 *
	 * @default false
	 */
	throwOnError?: T['throwOnError']
}
// ...
export interface ClientOptions {
	baseURL?: string
	throwOnError?: boolean
}